Create file watchtyped.go (split watch.go to 2)
